Conflicts:
	core/fixel/helpers.h
	core/fixel/types.h
	src/fixel/matrix.cpp
	src/fixel/matrix.h

+++ src/registration/multi_resolution_lmax.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2008-2019 the MRtrix3 contributors.
+/* Copyright (c) 2008-2023 the MRtrix3 contributors.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -31,10 +31,10 @@ namespace MR
     FORCE_INLINE ImageType multi_resolution_lmax (ImageType& input,
                                                   const default_type scale_factor,
                                                   const bool do_reorientation = false,
-                                                  const int lmax = 0)
+                                                  const uint32_t lmax = 0)
     {
-      vector<int> from (input.ndim(), 0);
-      vector<int> size (input.ndim());
+      vector<uint32_t> from (input.ndim(), 0);
+      vector<uint32_t> size (input.ndim());
       for (size_t dim = 0; dim < input.ndim(); ++dim)
         size[dim] = input.size(dim);
       if (do_reorientation)
@@ -63,13 +63,13 @@ namespace MR
                                                   const vector<MultiContrastSetting>& contrast,
                                                   vector<MultiContrastSetting>* contrast_updated = nullptr)
     {
-      vector<int> volume_indices;
+      vector<uint32_t> volume_indices;
       size_t start = 0;
       for (size_t ic = 0; ic < contrast.size(); ic++) {
         const auto & mc = contrast[ic];
         assert (mc.nvols > 0);
         for (size_t i = 0; i < mc.nvols; i++)
-          volume_indices.push_back((int) (mc.start + i));
+          volume_indices.push_back(mc.start + i);
         // adjust start index to be relative to subset
         if (contrast_updated)
           (*contrast_updated)[ic].start = start;
